[ADD] new default value for the RSID displayed for the user
This new value (named DISPLAY_RSID) will be used to display information related at the GPS coordinates for the final user (for example in the detail views or in the pdfs or other generated documents).
Like all other parameters, it can be override by putting a new value in the custom.py file.

It comes in addition with the values RSID which is used for the formatting of geometric values in database and the value API_RSID which is used to interact with the map and leaflet.