Fix failure of E2E test suite due to memcached connection
This change fixes the run of the E2E test suite inside Github Actions jobs testing the .deb packages for Ubuntu 18.04 and 20.04.

This issue was not reproductible locally on a vbox VM where the E2E test suite runs with no error.

Exceptions caught in an adhoc sentry project showed that the application could not connect to its cache backend, which is memcached. This follows the upgrade to django 3.2 and the change to pymemcache package for the memcached backend (as recommended in Django's doc: https://docs.djangoproject.com/en/3.2/releases/3.2/#pymemcache-support and https://docs.djangoproject.com/en/3.2/topics/cache/#memcached).