Fix #3494: nearby sensitive areas now appears in outdoor details pages (and the other way too)
The `SensitiveArea` model missed the `sites` and `courses` properties. And `Site` and `Course` missed the `sensitive_areas` property.

Since it is template mechanism there are no auto tests on this + the `valuelist` mapentity tags silently dealt with the missing properties.

The following properties are added too for consistency with all other models:
- `SensitiveArea.published_courses`
- `SensitiveArea.published_sites`
- `Site.published_sensitive_areas`
- `Course.published_sensitive_areas`